shop

+++ FlexTablesDeletePost.php
@@ -7,8 +7,9 @@ class FlexTablesDeletePost extends ValidationId{
     private $modal;
     function __construct(){
         $this->modal = new ModelJson($_GET['id']);
-        parent::__construct($this->getMyModal()); 
-        $this->getMyModal()->saveModel($this->deleteItem($_GET['id'], $this->getMyModal()->getObj()));
+        parent::__construct($this->getMyModal());
+        if(!isset($_POST['Branches']) && !isset($_POST['choices']))
+            $this->getMyModal()->saveModel($this->deleteItem($_GET['id'], $this->getMyModal()->getObj()));
         MyFlexTablesView::initMyFlexTablesView('Delete');
     }
     function getMyModal(){
